requirements.txt
# Core dependencies
openai>=1.0.0
requests>=2.25.0
python-dotenv>=0.15.0
beautifulsoup4>=4.9.0
tiktoken>=0.3.0

# Optional dependencies
pytest>=6.0.0
black>=21.5b2
tqdm>=4.64.0